requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
